test: 시그니처 생성 테스트 추가 #65
